Split settings into “general” and “”java
